Black and white as well as move to python.
